ci: Add build dir as an artifact to keep static analysis working
Currently GitLab CI uploads the whole build directory as an artifact,
configured in the shared ci-before.yml in ci-tooling. This is necessary
for the static analysis job to work. However, it's very wasteful for the
majority of repositories that don't use the static analysis job, so we'll
soon remove that from the shared config.

This commit explicitly adds the build dir to the list of artifacts to
ensure that the static analysis job continues to work in Keysmith after
the shared CI config is changed.